Moved collision rate functions from parameters.py to collisions.py
