+++ plasmapy/physics/transport/__init__.py
@@ -1,5 +1,7 @@
 from .collisions import (Coulomb_logarithm,
                          collision_frequency,
+                         collision_rate_electron_ion,
+                         collision_rate_ion_ion,
                          b_perp,
                          impact_parameter,
                          mean_free_path,
